RAG_GuardRails_Evals/Assignment-Guardrails-Evals/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Document processing
docling==2.15.0
python-docx==1.1.2
PyMuPDF==1.24.14

# Vector store
qdrant-client==1.13.1

# LLM & Embeddings
groq==0.13.1
openai==1.59.9        # used for embeddings only (text-embedding-3-small)
tiktoken==0.8.0

# Semantic routing
semantic-router==0.1.3

# PII detection
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355
spacy==3.8.3

# Evaluation
ragas==0.2.9
datasets==3.2.0

# Utilities
python-dotenv==1.0.1
pandas==2.2.3
httpx==0.28.1
aiofiles==24.1.0
tenacity==9.0.0

